move model conversions to api package
